deepen: inline x-sentiment-eligibility into X Sentiment provider
The eligibility gate (form importance + ticker normalization + mega-cap
allowlist) sat in a 146-LOC module with a single production caller —
`getXSentiment` in `x-sentiment-provider.ts`. Tests on the shallow module
asserted on internal branches that are already observable through the
provider's `skipReason` field, so the interface was nearly as complex as
the implementation.

Move all eligibility logic (constants, allowlist cache, `checkEligibility`,
`normalizeTicker`) into `x-sentiment-provider.ts` as private implementation.
Re-export `EligibilityReason` and expose `_internal.checkEligibility` so the
G2 eval runner (an internal seam used by the integration eval) can still
reach the gate without re-implementing it.

The deletion test passes: complexity concentrates inside the provider,
where it belongs; callers see one entry point (`getXSentiment`) and one
result type. The interface is the test surface — eligibility decisions
are now observed through `result.skipReason`.

Tests: delete `x-sentiment-eligibility.test.ts` (24 direct-module tests,
testing past the new interface). Add 6 tests at the provider interface
covering invalid ticker format, unknown form type, ticker normalization,
case-insensitive form match, and the `X_SENTIMENT_ALLOWLIST_EXTRA` env
override. 30 provider tests pass; integration eval (98 entries) passes.

Add `X Sentiment` term to CONTEXT.md.

Dependency category: in-process (no I/O behind the gate). No adapter
needed — the seam is internal to the provider.
