+++ lib/ai/x-sentiment-provider.ts
@@ -33,7 +33,6 @@ import {
   recordEnrichmentFailureNoRefund,
 } from './enrichment-spend';
 import type { EnvelopeKind } from '@prisma/client';
-import { checkXSentimentEligibility, type EligibilityReason } from './x-sentiment-eligibility';
 import {
   validateXSentiment,
   type XSentiment,
@@ -54,6 +53,113 @@ const DEFAULT_WINDOW_HOURS = 168; // 7 days — captures pre-filing speculation 
 const MAX_LABEL_LENGTH = 200;
 const MAX_CONTEXT_LENGTH = 500;
 const APPROX_COST_PER_CALL_USD = 0.05;
+
+// Eligibility gate combines three filters before the (expensive) x_search call
+// fires:
+//   1. Form importance — high-impact form types only (D4=B salience targeting)
+//   2. Strict ticker normalization — exact canonical match, no fuzzy (F2 cashtag collision)
+//   3. Mega-cap allowlist — defense against pump-and-dump on low-float tickers (F2 + F5)
+//
+// Forms set is inlined (rather than read from `form-registry.ts`) because the
+// registry has a stale `.js` shadow that wins jest module resolution.
+const TICKER_REGEX = /^[A-Z]{1,5}(?:\.[A-Z]{1,2})?$/;
+
+// Mirrors the `importance: 'high'` set in `lib/sec-edgar/form-registry.ts`.
+const HIGH_IMPORTANCE_FORMS: ReadonlySet<string> = new Set<string>([
+  '10-K', '10-K/A',
+  '10-Q', '10-Q/A',
+  '8-K',
+  '20-F', '40-F',
+  'FORM4', '4',
+  'FORM 144', '144',
+  'SC 13D', 'SC13D',
+  'DEF 14A',
+  'S-1', 'S-3', 'S-4',
+  'F-1', 'F-3', 'F-4',
+]);
+
+// Forms we recognize but treat as low-signal for sentiment. Used to
+// distinguish "unknown form" from "known but not eligible" so the skip
+// reason is precise.
+const KNOWN_LOW_IMPORTANCE_FORMS: ReadonlySet<string> = new Set<string>([
+  'NT 10-K', 'NT 10-Q',
+  '10-K/A_NT', '10-Q/A_NT',
+  '11-K', '15', '15-12B', '15-12G',
+  'SD', 'CORRESP', 'UPLOAD',
+  'SC 13G', 'SC13G',
+  '3', 'FORM3', '5', 'FORM5',
+]);
+
+// Curated mega-cap allowlist (~S&P 100). Sorted alphabetically for diff
+// readability. Extend at runtime via `X_SENTIMENT_ALLOWLIST_EXTRA` env var.
+const ALLOWLIST_BASE: readonly string[] = [
+  'AAPL', 'ABBV', 'ABT', 'ACN', 'ADBE', 'AIG', 'AMD', 'AMGN', 'AMT', 'AMZN',
+  'AVGO', 'AXP', 'BA', 'BAC', 'BIIB', 'BK', 'BKNG', 'BLK', 'BMY', 'BRK.A',
+  'BRK.B', 'C', 'CAT', 'CHTR', 'CL', 'CMCSA', 'COF', 'COP', 'COST', 'CRM',
+  'CSCO', 'CVS', 'CVX', 'DD', 'DE', 'DHR', 'DIS', 'DOW', 'DUK', 'EMR',
+  'EXC', 'F', 'FDX', 'GD', 'GE', 'GILD', 'GM', 'GOOG', 'GOOGL', 'GS',
+  'HD', 'HON', 'IBM', 'INTC', 'ISRG', 'JNJ', 'JPM', 'KHC', 'KMI', 'KO',
+  'LIN', 'LLY', 'LMT', 'LOW', 'MA', 'MCD', 'MDLZ', 'MDT', 'MET', 'META',
+  'MMM', 'MO', 'MRK', 'MS', 'MSFT', 'NEE', 'NFLX', 'NKE', 'NVDA', 'ORCL',
+  'PEP', 'PFE', 'PG', 'PM', 'PYPL', 'QCOM', 'RTX', 'SBUX', 'SCHW', 'SO',
+  'SPG', 'T', 'TGT', 'TMO', 'TMUS', 'TSLA', 'TXN', 'UNH', 'UNP', 'UPS',
+  'USB', 'V', 'VZ', 'WBA', 'WFC', 'WMT', 'XOM',
+];
+
+let cachedAllowlist: Set<string> | null = null;
+
+function getAllowlist(): Set<string> {
+  if (cachedAllowlist) return cachedAllowlist;
+  const extra = (process.env.X_SENTIMENT_ALLOWLIST_EXTRA ?? '')
+    .split(',')
+    .map((s) => s.trim().toUpperCase())
+    .filter(Boolean);
+  cachedAllowlist = new Set([...ALLOWLIST_BASE, ...extra]);
+  return cachedAllowlist;
+}
+
+function normalizeTicker(input: string | null | undefined): string {
+  if (!input) return '';
+  return input.trim().toUpperCase().replace(/^\$+/, '');
+}
+
+export type EligibilityReason =
+  | 'no_ticker'
+  | 'invalid_ticker_format'
+  | 'unknown_form_type'
+  | 'low_importance_form'
+  | 'not_in_allowlist';
+
+interface EligibilityResult {
+  eligible: boolean;
+  ticker: string;
+  reason?: EligibilityReason;
+}
+
+function checkEligibility(ticker: string | null | undefined, formType: string | null | undefined): EligibilityResult {
+  const normalized = normalizeTicker(ticker);
+  if (!normalized) {
+    return { eligible: false, ticker: normalized, reason: 'no_ticker' };
+  }
+  if (!TICKER_REGEX.test(normalized)) {
+    return { eligible: false, ticker: normalized, reason: 'invalid_ticker_format' };
+  }
+  const form = String(formType ?? '').toUpperCase().trim();
+  if (!form) {
+    return { eligible: false, ticker: normalized, reason: 'unknown_form_type' };
+  }
+  if (HIGH_IMPORTANCE_FORMS.has(form)) {
+    // fall through — eligible on form dimension
+  } else if (KNOWN_LOW_IMPORTANCE_FORMS.has(form)) {
+    return { eligible: false, ticker: normalized, reason: 'low_importance_form' };
+  } else {
+    return { eligible: false, ticker: normalized, reason: 'unknown_form_type' };
+  }
+  if (!getAllowlist().has(normalized)) {
+    return { eligible: false, ticker: normalized, reason: 'not_in_allowlist' };
+  }
+  return { eligible: true, ticker: normalized };
+}
 
 export interface XSentimentInput {
   ticker: string;
@@ -196,10 +302,7 @@ function buildContext(s: XSentiment): string {
  */
 export async function getXSentiment(input: XSentimentInput): Promise<XSentimentResult> {
   const accessionNumber = input.accessionNumber;
-  const eligibility = checkXSentimentEligibility({
-    ticker: input.ticker,
-    formType: input.formType,
-  });
+  const eligibility = checkEligibility(input.ticker, input.formType);
   if (!eligibility.eligible) {
     monitoring.incrementCounter(`ai.${PROVIDER_NAME}_skipped_${eligibility.reason}`, 1);
     return { enrichment: null, skipReason: eligibility.reason };
@@ -381,6 +484,11 @@ export async function getXSentiment(input: XSentimentInput): Promise<XSentimentR
   };
 }
 
+/** Test-only: reset env-derived allowlist cache between tests. */
+export function _resetAllowlistCache(): void {
+  cachedAllowlist = null;
+}
+
 export const _internal = {
   PROVIDER_NAME,
   SECTION_HEADER,
@@ -394,4 +502,11 @@ export const _internal = {
   tryParseModelJson,
   buildLabel,
   buildContext,
+  TICKER_REGEX,
+  HIGH_IMPORTANCE_FORMS,
+  KNOWN_LOW_IMPORTANCE_FORMS,
+  ALLOWLIST_BASE,
+  getAllowlist,
+  normalizeTicker,
+  checkEligibility,
 };
